feat: define types for createFloor
